requirements.txt
asgiref==3.9.1
astroid==3.3.10
autopep8==2.0.4
black==25.1.0
click==8.2.1
dill==0.4.0
Django==5.2.4
django-stubs==5.2.1
django-stubs-ext==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docstring-to-markdown==0.17
flake8==7.1.2
importlib_metadata==8.7.0
isort==6.0.1
jedi==0.19.2
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.9.0
pylint==3.3.7
pylsp-mypy==0.7.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.13.0
pytoolconfig==1.3.1
rope==1.13.0
snowballstemmer==3.0.1
sqlparse==0.5.3
tomlkit==0.13.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.1
ujson==5.10.0
whatthepatch==1.0.7
yapf==0.43.0
zipp==3.23.0
